requirements.txt
pytest==8.3.3
pytest-html==4.1.1
pytest-xdist==3.6.1
PyYAML==6.0.2
Pillow==10.4.0
numpy==1.26.4
pytesseract==0.3.13
pandas==2.2.3
opencv-python==4.8.1.78
robotframework==6.1.1
Appium-Python-Client==3.1.1
schedule==1.2.0
openpyxl==3.1.2